ApPStorage add

+++ Utterance/UI/Controls/PitchStepper.swift
@@ -8,10 +8,10 @@
 import SwiftUI
 
 struct PitchStepper: View {
-    @Binding var pitchValue: Float
+    @Binding var pitchValue: Double
 
-    private let range: ClosedRange<Float> = 0.5...2.0
-    private let step: Float = 0.5
+    private let range: ClosedRange<Double> = 0.5...2.0
+    private let step: Double = 0.5
     
     var body: some View {
         Stepper("Pitch: \(String(format: "%.1f", pitchValue))",
@@ -29,7 +29,7 @@ struct PitchStepper: View {
 
 
 struct PitchStepper_Previews: PreviewProvider {
-    @State static var pitchValue: Float = 1
+    @State static var pitchValue: Double = 1
     
     static var previews: some View {
         PitchStepper(pitchValue: $pitchValue)
